renamed "misc" to "main" as it contains main functions
